requirements.txt
gevent==20.9.0
gunicorn==20.0.4
tensorflow==2.2.0
pillow~=7.2.0
numpy==1.18.1
urllib3==1.25.9
flask==1.1.2
Flask-PyMongo==2.3.0
smartcrop==0.3.2